Separate DOM translation functions and DOM overlays
